+++ bootstrap.cc
@@ -1,0 +1,76 @@
+#include <bits/stdc++.h>
+#include <boost/geometry.hpp>
+#include <boost/geometry/geometries/point.hpp>
+#include <boost/geometry/geometries/box.hpp>
+#include <boost/geometry/index/rtree.hpp>
+#define MAX 20000
+#define rep(i,n) for(int i = 0; i < (n); ++i)
+#define drep(i,n) for(int i = (n)-1; i >= 0; --i)
+#define srep(i,s,t) for (int i = s; i < t; ++i)
+#define rng(a) a.begin(),a.end()
+#define sz(x) (int)(x).size()
+#define show(x) cout<<#x<<" = "<<x<<endl;
+#define chmin(x,y) x=min(x,y)
+#define chmax(x,y) x=max(x,y)
+using namespace std;
+typedef vector<int> vi;
+typedef vector<vi> vvi;
+const int INF = 1001001001;
+
+namespace bg = boost::geometry;
+namespace bgi = boost::geometry::index;
+typedef bg::model::point<float, 2, bg::cs::cartesian> point;
+
+extern int n, length, tour[MAX];
+extern float city[MAX][2];
+
+vi neighbor[MAX];
+
+bgi::rtree<pair<point,unsigned>,bgi::quadratic<MAX>> rtree;
+typedef vector<pair<point,unsigned>> vp;
+
+int dist(int i, int j) {
+  float xd = city[i][0] - city[j][0];
+  float yd = city[i][1] - city[j][1];
+  return (int)(sqrt(xd * xd + yd * yd) + .5);
+}
+
+int xor64() {
+  static uint64_t x = 88172645463325252ULL;
+  x = x ^ (x << 13); x = x ^ (x >> 7);
+  x = x ^ (x << 17);
+  return abs(int(x));
+}
+
+int bestlen=INF;
+vi bestlog;
+
+void snapshot() {
+  if(bestlen>length) {
+    chmin(bestlen,length);
+    bestlog.resize(n);
+    rep(i,n) {
+      bestlog[i]=tour[i];
+    }
+  }
+}
+
+
+void bootstrap() {
+  // build rtree
+  rep(i,n) {
+    point p(city[i][0],city[i][1]);
+    rtree.insert(make_pair(p,i));
+  }
+  // build neighbor
+  rep(i,n) {
+    point p(city[i][0],city[i][1]);
+    vp dst;
+    rtree.query(bgi::nearest(p,int(10*log2(n))),back_inserter(dst));
+    for(auto nea:dst) {
+      int id = nea.second;
+      if(id==i) continue;
+      neighbor[i].push_back(id);
+    }
+  }
+}
