+++ heuristics.cc
@@ -1,0 +1,10 @@
+extern void bootstrap();
+extern void build();
+extern void run();
+
+int tspSolver() {
+  bootstrap();
+  build();
+  run();
+  return 1;
+}
